DCGAN_anime/requirements.txt
torchvision==0.8.1+cu110
torch==1.7.0+cu110
Pillow==8.0.1
